Created MockNotificationsService to fix tests
