+++ unstoppable-ios-app/domains-manager-ios/AppContext/MockContext.swift
@@ -15,12 +15,7 @@ final class MockContext: AppContextProtocol {
                                                                                                        walletConnectServiceV2: walletConnectServiceV2, walletConnectRequestsHandlingService: wcRequestsHandlingService)
     private(set) lazy var imageLoadingService: ImageLoadingServiceProtocol = ImageLoadingService(qrCodeService: qrCodeService)
     private(set) lazy var networkReachabilityService: NetworkReachabilityServiceProtocol? = NetworkReachabilityService()
-    private(set) lazy var notificationsService: NotificationsServiceProtocol = {
-        NotificationsService(externalEventsService: externalEventsService,
-                             permissionsService: permissionsService,
-                             udWalletsService: udWalletsService,
-                             wcRequestsHandlingService: wcRequestsHandlingService)
-    }()
+    private(set) lazy var notificationsService: NotificationsServiceProtocol = { MockNotificationsService() }()
     private(set) lazy var permissionsService: PermissionsServiceProtocol = PermissionsService()
     private(set) lazy var pullUpViewService: PullUpViewServiceProtocol = PullUpViewService(authentificationService: authentificationService)
     private(set) lazy var toastMessageService: ToastMessageServiceProtocol = ToastMessageService()
